+++ src/Data/KindID/Unsafe.hs
@@ -1,0 +1,11 @@
+module Data.KindID.Unsafe
+  (
+  -- * Unsafe 'KindID' decoding
+    unsafeParseString
+  , unsafeParseText
+    , unsafeParseByteString
+  -- * Unsafe conversion
+    , unsafeFromTypeID
+  ) where
+
+import           Data.KindID.Internal